Refactor CreditRequest and CreditRequestItems classes to use Lombok annotations for boilerplate code reduction
- Added @Getter, @Setter, @NoArgsConstructor, @AllArgsConstructor, and @Builder annotations to CreditRequest and CreditRequestItems classes.
- Improved readability and maintainability by reducing boilerplate code.
- Updated CreditRequestValidationService to streamline validation logic and improve exception handling.
- Enhanced CreateCreditRequestCase to utilize context-based validation for items, reducing database queries and improving performance.
- Refactored item processing logic to leverage pre-loaded validation contexts, ensuring efficient validation and processing of credit requests.

+++ src/main/java/br/sptrans/scd/creditrequest/application/port/in/CreditRequestManagementUseCase.java
@@ -69,10 +69,10 @@ public interface CreditRequestManagementUseCase {
             String codCanal,
             String codProduto,
             String codSituacao,
-            Double vlItem,
-            Double vlTxadm,
-            Double vlTxserv
-            ) {
+            BigDecimal vlItem,
+            BigDecimal vlTxadm,
+            BigDecimal vlTxserv
+    ) {
 
     }
